fix: tests on Order Model (user fk)

+++ migrations/sqls/20230121060858-orders-table-up.sql
@@ -1,5 +1,5 @@
 CREATE TABLE orders (
     id SERIAL PRIMARY  KEY,
     status VARCHAR,
-    user_id bigint REFERENCES users(id)
+    user_id int REFERENCES users(id)
 );
